refactor: extract meta content / def into their own files
