+++ projects/ngx-meta/src/core/src/ngx-meta-meta-content.ts
@@ -1,0 +1,10 @@
+/**
+ * Content to be set for a specific `<meta>` element in the page
+ *
+ * Can be `undefined` or `null`. In that case, the element will be removed.
+ *
+ * See {@link NgxMetaMetaService}
+ *
+ * @public
+ */
+export type NgxMetaMetaContent = string | undefined | null
